-renamed Avatar.tsx to CustomAvatar.tsx
