+++ docs/src/prototypes/hexagonalAvatar/CustomAvatar.tsx
